Split MinmodType into new file
