+++ src/Evolution/DiscontinuousGalerkin/SlopeLimiters/CMakeLists.txt
@@ -5,6 +5,7 @@ set(LIBRARY SlopeLimiters)
 
 set(LIBRARY_SOURCES
   Minmod.cpp
+  MinmodType.cpp
   )
 
 add_spectre_library(${LIBRARY} ${LIBRARY_SOURCES})
