+++ src/Evolution/DiscontinuousGalerkin/SlopeLimiters/Minmod.hpp
@@ -21,6 +21,7 @@
 #include "Domain/Element.hpp"  // IWYU pragma: keep
 #include "Domain/MaxNumberOfNeighbors.hpp"
 #include "ErrorHandling/Assert.hpp"
+#include "Evolution/DiscontinuousGalerkin/SlopeLimiters/MinmodType.hpp"
 #include "NumericalAlgorithms/LinearOperators/MeanValue.hpp"
 #include "Options/Options.hpp"
 #include "Utilities/Gsl.hpp"
@@ -60,14 +61,6 @@ template <size_t VolumeDim>
 struct SizeOfElement;
 }  // namespace Tags
 /// \endcond
-
-namespace SlopeLimiters {
-/// \ingroup SlopeLimitersGroup
-/// \brief Possible types of the minmod slope limiter.
-///
-/// \see SlopeLimiters::Minmod
-enum class MinmodType { LambdaPi1, LambdaPiN, Muscl };
-}  // namespace SlopeLimiters
 
 namespace Minmod_detail {
 // Encodes the return status of the minmod_tvbm function.
@@ -472,15 +465,3 @@ bool operator!=(const Minmod<VolumeDim, TagList>& lhs,
 }
 
 }  // namespace SlopeLimiters
-
-template <>
-struct create_from_yaml<SlopeLimiters::MinmodType> {
-  template <typename Metavariables>
-  static SlopeLimiters::MinmodType create(const Option& options) {
-    return create<void>(options);
-  }
-};
-template <>
-SlopeLimiters::MinmodType
-create_from_yaml<SlopeLimiters::MinmodType>::create<void>(
-    const Option& options);
